+++ convertionAPI/src/main/java/com/senai/convertionapi/dtos/RespostaDTO.java
@@ -1,24 +1,26 @@
 package com.senai.convertionapi.dtos;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
+import lombok.Getter;
+import lombok.Setter;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class RespostaDTO {
 
-    private Integer valorOriginal;
+    private Double valorOriginal;
     private String tipoConversao;
-    private Integer valorConvertido;
+    private Double valorConvertido;
     private String erro;
 
     public RespostaDTO() {
-        valorConvertido = 0;
+        valorConvertido = 0.0;
     }
 
-    public Integer getValorOriginal() {
+    public Double getValorOriginal() {
         return valorOriginal;
     }
 
-    public void setValorOriginal(Integer valorOriginal) {
+    public void setValorOriginal(Double valorOriginal) {
         this.valorOriginal = valorOriginal;
     }
 
@@ -30,11 +32,11 @@ public class RespostaDTO {
         this.tipoConversao = tipoConversao;
     }
 
-    public Integer getValorConvertido() {
+    public Double getValorConvertido() {
         return valorConvertido;
     }
 
-    public void setValorConvertido(Integer valorConvertido) {
+    public void setValorConvertido(Double valorConvertido) {
         this.valorConvertido = valorConvertido;
     }
